Adjust measure_angle + length detemination, cleanup
In order to use the new thinning function properly, the main function
had to be clean up and DnaBound and DnaFree had to be changed. But the
new thinning function still makes problems.

Also, the function to recover pixels lost in the thinning process has
been modularized and adjusted so that it can be used by Dennis and
Philips length determination function.